Introduce /tests with some dry run magic to check rollout status
This adds scripts that run ansible in "dry run" mode, parses
the output and displays them in human readable outputs. The
scripts are currently written very ugly and only as POC. However
later this could help to build something that automatically
informs us about necessary changes, which we should better roll
out. Or when someone has done manual changes to the system.