mesh_fastd_remotes_peers_git: track peers' last access time
To get an idea of which keys are still in use this commit introduces a cronjob which uses the status sockets of the fastd instances to retrieve the currently connected peers and touches (updates the last access time of) the files corresponding to the peers' names in the newly created directory '/var/fastd_peers'.
Due to privacy reasons this cronjob is being executed every 11 hours only.

[Edit-By lemoer: make crontab a one-liner, so cron likes it]