Implement experimental_supernodes variable
By default experimental supernodes are not rolled out. The newly
introduced variable "experimental_supernodes" can be used to roll them.

The new variable can be specified by adding the -e switch to your
ansible-playbook call:

    ansible-playbook -e experimental_supernodes=sn06:sn07 ...

This command will roll the supernodes playbook also on the
experimental supernodes sn06 and sn07. Multiple supernodes
are separated by ":".

Note that the experimental supernodes are added to the
inventory, but are not part of the "supernodes" ansible group.