treewide: only hide secret diffs for tests/run.sh
Before this commit, we redacted the diffs of some actions, because the
actions contained secret information. This was done, since we plan to send
the output tests/run.sh via mail in the future. However, the approach was
cumbersome, since we often need to review the changes, if we invoke
ansible manually (not via tests/run.sh).

Therefore this commit changes the behaviour, so the diffs are only
redacted, if ansible is invoked via tests/run.sh.