mesh_batman: explicitly set macs for batX on supernodes
Otherwise batman would use macs based on the order of the lower
interfaces being added to batman. Before using systemd-networkd
to add interfaces to batman, we had a more or less stable order.
Now, we use systemd-networkd and the order is kind of radom. So
in order to have a stable mac on batX devices, we set the macs
on batX explicit now.